Custom bs58check through base.js
